Connector for FreePBX
Added a sample connector to FreePBX system. It only supports
getting the extension's SIP credentials and getting the
list of other extensions.

While doing this ended up re-factoring some of the abstract connector class
and moving the functionality related to working with sessions
(if they are maintained on the BSS connector side) into a
separate module, so they can be re-used.