Updated API schema to v1.1.0
This includes adding the "custom" methods and ability
login a user into the app based on a click in SMS/email
or QR code using the "auto-provision" method.

An important change is that the "code" attribute in the error
response is now optional and should be provided in cases where
it is needed.